fpf_func_test: append white after fixed option
Currently, fpf_func_test combines few string to make an option for fpf.
And, because a string for fixed option that comes before assertion
option is finished without following white space, assertion options all
have a space at start of them. However, it would be more efficient to
give a space after the fixed option once than giving a space to every
assertion options. It would save more bytes of code ;)  For the reason,
this change do that.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>