add simple python wrapper for remote_exps.exp
For some people, expect is not a familiar tool.  For the case, this
commit adds a new python wrapper for `remote_exps.exp`.  The python
wrapper works for one remote experiment.  It may be used inside
experiment specification file as `main`, `back`, `end` or any type of
task and be used as a input to familiar `run_exps.py`.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>