tblprc: append_column: hide row from generator
append_column() function receives name of new column and a function for
new value of the column generation.  However, the generator function
receives table's row directly.  Because the function should touch the
row directly, it is unflexible to change row struct for future.  This
commit changes the semantic to pass index of each row to generator and
the generator to generate new value for the column of the row of the
index only.  The generator function may access the table via ATable's
item_at() method.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>