tblprc: change table inner struct
This commit changes table inner struct for row from list of values to
dictionary.  The dictionary has column name as key and value as value.
However, the constructor of table is still receiving list of list as
rows as before.  It is intended to be easy for writing test code.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>