expect: Expect shell prompt more appropriately
`expect` scripts for remote machine experiments is expecting `hostname`
only to ensure termination of each command.  However, if the command put
output including the `hostname`, the expect script cannot guarantee the
waiting.  This commit work around the case by expecting `*$hostname*$ `
or `*$hostname*# ` for prompt.  Of course, this is just a trick that can
be fail to gauratee the behavior if the command put output including the
pattern.  However, for now, it would work for most case.

Setting `PS1` as hash of hostname and waiting for it would be one way to
do the job more correctly, though concrete solution is not ready yet.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>