perfscr-nrevt: Print time relatively
Currently, perfscr-nrevt prints count of events by time.  The time is
seconds in timestamp.  However, because the seconds in timestamp has no
big meaning, it just complicates output and confuse user.  This commit
modifies the time field to show relative time from start of the trace to
simplify the output and enhance readability.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>