ltldat/normalize(): Process basis cell at last
Because normalization calculates and overwrite cells in order, in case
of basis column is not a last column, the basis values become one and
then it is used to calculate normalized value of next columns.  It makes
no sense.  This commit fixes the problem by handling of the basis column
at last.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>