Add script for work waiting
Some tasks work in background.  Daemon processes such as mysqld can be
an example.  To wait the background job, monitoring its status is
necessary.  This commit adds a script for automation of the waiting.  It
monitor status of processes of a command and wait until all of the
processes becomes non-running status.

Signed-off-by: sjpark <sjpark@hydra.hydra>