splitplot: Receive record, not pinatrace output
splitplot is assuming the input file to be a pinatrace output.  It is
not proper action for its name and self description.  This commit
changes the tool to assume that the input is in single record format.
In other words, if you were using splitplot to plot your pinatrace
output, you should use `pinatrace2rec` explicitly.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>