lzhwinfo: Update output format
It would be more natural to print total number of hardware threads after
number of socks, number of cores per sock and number of threads per core
because number of socks * number of cores per sock * number of threads
per core = number of hardware threads.  This commit modifies the output
format to be in the order.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>