Add directory for CVE stat analysis
According to a maintainers summit topic discussion[1], 99% of security
issues are fixed in stable kernels _BEFORE_ they are known.  No doubt
about it, but having a tool for anyone to get the data on their own
would be nice.  Let's make it.  First, add the question that I want to
answer with the tool.

[1] https://lore.kernel.org/ksummit/2023081540-vindicate-caterer-33c6@gregkh/

Signed-off-by: sjpark <sj38.park@gmail.com>