cve_stat/cvestat_percentile: Support broken-to-committed, and add an option to skip negative values
Yet another interesting point would be how long it take for the CVE to
be reported after the broken commit is committed.  Also, it is useful to
have a way to skip negative values, for logscale plotting.  Do those.
Actually each change should be committed separately, but I forgot
committing the previous change.  Not a big deal, so just make the two
changes in this one commit.

Signed-off-by: sjpark <sj38.park@gmail.com>